use separate module for translator
