+++ modules/translator.py
@@ -1,0 +1,15 @@
+import translators
+from functools import lru_cache
+
+@lru_cache(maxsize=32, typed=False)
+def translate2en(text, element):
+    if not text:
+        return text
+
+    try:
+        result = translators.translate_text(text,to_language='en')
+        print(f'Translated {element}: {result}')
+        return result
+    except Exception as e:
+        print(f'Error during translation of {element}: {e}')
+        return text
